pyproject.toml
[project]
name = "alpha-q"
version = "0.1.0"
description = "Deep Q-Network variants for Atari — from vanilla DQN to Rainbow"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "torch>=2.0",
    "gymnasium[atari]>=0.29",
    "ale-py>=0.9",
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "mlflow>=2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.3",
]
notebooks = [
    "jupyter>=1.0",
    "matplotlib>=3.7",
]
video = [
    "moviepy>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.3",
]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alpha_q"]
